third_party/WebKit/LayoutTests/presentation/presentation-receiver.html

<!DOCTYPE html>
<html>
<body>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script>

test(t => {
  assert_true('receiver' in navigator.presentation);
}, "Test that presentation.receiver is present.")

test(t => {
  assert_equals(navigator.presentation.receiver, null);
}, "Test that presentation.receiver returns null when the setting is not set.");

test(t => {
  internals.settings.setPresentationReceiver(true);
  t.add_cleanup(_ => { internals.settings.setPresentationReceiver(false); });

  assert_not_equals(navigator.presentation.receiver, null);
}, "Test that presentation.receiver returns non-null when the setting is set.");

test(t => {
  internals.settings.setPresentationReceiver(true);
  t.add_cleanup(_ => { internals.settings.setPresentationReceiver(false); });

  assert_equals(typeof(navigator.presentation.receiver), "object");

  assert_true('connectionList' in navigator.presentation.receiver);
  assert_equals(typeof(navigator.presentation.receiver.connectionList), "object");
}, "Test the Presentation Receiver API properties types and presence.");

</script>
</body>
</html>
